+++ src/components/Chat/ChatMessages.tsx
@@ -6,7 +6,7 @@ import rehypeRaw from 'rehype-raw'
 import { FaRegUserCircle } from 'react-icons/fa'
 import '../../styles/chat.css'
 import { Chat, ChatFilters, ReorChatMessage } from './types'
-import { useTabsContext } from '@/contexts/TabContext'
+import { useTabsContext } from '@/contexts/WindowContentContext'
 import ChatInput from './ChatInput'
 import { getClassNameBasedOnMessageRole, getDisplayMessage } from './utils'
 import LoadingDots from '@/utils/animations'
